requirement.txt
pandas
openpyxl
faiss-cpu
sentence-transformers
flask
ollama
